Move test case.
